requirements.txt
Flask == 0.10.1
requests == 2.2.1
Werkzeug == 0.10.4
Pillow == 3.1.0